authentication and authorization implemented

+++ src/api/categories/handler.js
@@ -12,20 +12,27 @@ class CategoriesHandler {
   async postCategoryHandler(request, h) {
     this._validator.validateCategoryPayload(request.payload)
     const { name } = request.payload
+    const { id: credentialId } = request.auth.credentials
 
-    const category = await this._categoriesService.addCategory({ name })
+    const categoryId = await this._categoriesService.addCategory({
+      name,
+      owner: credentialId
+    })
 
     const response = h.response({
       status: 'success',
       message: 'Berhasil menambahkan kategori',
-      data: category
+      data: {
+        categoryId
+      }
     })
     response.code(201)
     return response
   }
 
-  async getCategoriesHandler() {
-    const categories = await this._categoriesService.getCategories()
+  async getCategoriesHandler(request) {
+    const { id: credentialId } = request.auth.credentials
+    const categories = await this._categoriesService.getCategories(credentialId)
 
     return {
       status: 'success',
@@ -37,7 +44,9 @@ class CategoriesHandler {
 
   async getCategoryByIdHandler(request, h) {
     const { categoryId } = request.params
+    const { id: credentialId } = request.auth.credentials
 
+    await this._categoriesService.verifyCategoryOwner(categoryId, credentialId)
     const category = await this._categoriesService.getCategoryById(categoryId)
     const product = await this._productsService.getProducts({ categoryId })
     return {
@@ -54,7 +63,9 @@ class CategoriesHandler {
   async putCategoryByIdHandler(request) {
     const { categoryId } = request.params
     const { name } = request.payload
+    const { id: credentialId } = request.auth.credentials
 
+    await this._categoriesService.verifyCategoryOwner(categoryId, credentialId)
     await this._categoriesService.updateCategoryById(categoryId, { name })
 
     return {
@@ -65,7 +76,9 @@ class CategoriesHandler {
 
   async deleteCategoryByIdHandler(request) {
     const { categoryId } = request.params
+    const { id: credentialId } = request.auth.credentials
 
+    await this._categoriesService.verifyCategoryOwner(categoryId, credentialId)
     await this._categoriesService.deleteCategoryById(categoryId)
 
     return {
